Move decoder -> legacy_decoder_2
Update internal stage files and kafkaapi response headers to use
legacy_decoder_2 instead of decoder. This change ensures consistent
use of the updated legacy decoder, improving compatibility and
maintainability across the codebase. Removed deprecated decoder imports
to reduce confusion and streamline dependency management.
