+++ protocol/legacy_decoder_2/inspectable_hex_dump_test.go
